backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.2
opencv-python==4.8.1.78
pillow==10.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
numpy==1.24.3
bcrypt==4.1.2
python-magic==0.4.27
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
alembic==1.12.1